requirements/requirements.txt
# pygtk needs to be installed beforehand using the OS + linked into the virtualenv.
# rsvg  can be difficult to install + is only needed for examples that read SVG.
Pillow==1.7.8
pymetar==0.19    # examples/demos/pymetar.bot
numpy==1.6.2
pycairo
pygtk            # Used in window mode
rsvg
